+++ docs/conf.py
@@ -42,11 +42,15 @@ extensions = [
     "sphinx.ext.doctest",
     "numpydoc",
     "sphinx_autosummary_accessors",
+    "nbsphinx",
+    "sphinx.ext.mathjax",
 ]
 
 autosummary_generate = True
 # numpydoc_class_members_toctree = True
 numpydoc_show_class_members = False
+nbsphinx_kernel_name = "python3"
+nbsphinx_execute = "always"
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]