+++ rigid_body_motion/__init__.py
@@ -1,12 +1,12 @@
 """Top-level package for rigid-body-motion."""
 __author__ = """Peter Hausamann"""
 __email__ = "peter.hausamann@tum.de"
-__version__ = "0.4.1"
+__version__ = "0.5.0"
 from pathlib import Path
 
 from pkg_resources import resource_filename
 
-from . import ros  # noqa
+from . import plot, ros  # noqa
 from .coordinate_systems import (
     cartesian_to_polar,
     cartesian_to_spherical,
@@ -72,6 +72,7 @@ __all__ = [
     "best_fit_transform",
     "iterative_closest_point",
     "lookup_transform",
+    "lookup_pose",
     "lookup_twist",
     "lookup_linear_velocity",
     "lookup_angular_velocity",
@@ -106,7 +107,14 @@ example_data = {
 
 
 def transform_vectors(
-    arr, into, outof=None, dim=None, axis=None, timestamps=None, time_axis=None
+    arr,
+    into,
+    outof=None,
+    dim=None,
+    axis=None,
+    timestamps=None,
+    time_axis=None,
+    return_timestamps=False,
 ):
     """ Transform an array of vectors between reference frames.
 
@@ -144,6 +152,9 @@ def transform_vectors(
         The axis of the array representing the timestamps of the vectors.
         Defaults to the first axis of the array.
 
+    return_timestamps: bool, default False
+        If True, also return the timestamps after the transformation.
+
     Returns
     -------
     arr_transformed: array_like
@@ -157,12 +168,27 @@ def transform_vectors(
     transform_quaternions, transform_points, ReferenceFrame
     """
     return _transform(
-        "transform_vectors", arr, into, outof, dim, axis, timestamps, time_axis
+        "transform_vectors",
+        arr,
+        into,
+        outof,
+        dim,
+        axis,
+        timestamps,
+        time_axis,
+        return_timestamps=return_timestamps,
     )
 
 
 def transform_points(
-    arr, into, outof=None, dim=None, axis=None, timestamps=None, time_axis=None
+    arr,
+    into,
+    outof=None,
+    dim=None,
+    axis=None,
+    timestamps=None,
+    time_axis=None,
+    return_timestamps=False,
 ):
     """ Transform an array of points between reference frames.
 
@@ -176,9 +202,9 @@ def transform_points(
         which the array will be represented after the transformation.
 
     outof: str or ReferenceFrame, optional
-        ReferenceFrame instance or name of a registered reference frame in
-        which the array is currently represented. Can be omitted if the array
-        is a DataArray whose ``attrs`` contain a "representation_frame" entry
+        ReferenceFrame instance or name of a registered reference frame which
+        is the current reference frame of the array. Can be omitted if the
+        array is a DataArray whose ``attrs`` contain a "reference_frame" entry
         with the name of a registered frame.
 
     dim: str, optional
@@ -200,6 +226,9 @@ def transform_points(
         The axis of the array representing the timestamps of the points.
         Defaults to the first axis of the array.
 
+    return_timestamps: bool, default False
+        If True, also return the timestamps after the transformation.
+
     Returns
     -------
     arr_transformed: array_like
@@ -213,12 +242,27 @@ def transform_points(
     transform_vectors, transform_quaternions, ReferenceFrame
     """
     return _transform(
-        "transform_points", arr, into, outof, dim, axis, timestamps, time_axis
+        "transform_points",
+        arr,
+        into,
+        outof,
+        dim,
+        axis,
+        timestamps,
+        time_axis,
+        return_timestamps=return_timestamps,
     )
 
 
 def transform_quaternions(
-    arr, into, outof=None, dim=None, axis=None, timestamps=None, time_axis=None
+    arr,
+    into,
+    outof=None,
+    dim=None,
+    axis=None,
+    timestamps=None,
+    time_axis=None,
+    return_timestamps=False,
 ):
     """ Transform an array of quaternions between reference frames.
 
@@ -232,9 +276,9 @@ def transform_quaternions(
         which the array will be represented after the transformation.
 
     outof: str or ReferenceFrame, optional
-        ReferenceFrame instance or name of a registered reference frame in
-        which the array is currently represented. Can be omitted if the array
-        is a DataArray whose ``attrs`` contain a "representation_frame" entry
+        ReferenceFrame instance or name of a registered reference frame which
+        is the current reference frame of the array. Can be omitted if the
+        array is a DataArray whose ``attrs`` contain a "reference_frame" entry
         with the name of a registered frame.
 
     dim: str, optional
@@ -255,6 +299,9 @@ def transform_quaternions(
     time_axis: int, optional
         The axis of the array representing the timestamps of the quaternions.
         Defaults to the first axis of the array.
+
+    return_timestamps: bool, default False
+        If True, also return the timestamps after the transformation.
 
     Returns
     -------
@@ -277,6 +324,7 @@ def transform_quaternions(
         axis,
         timestamps,
         time_axis,
+        return_timestamps=return_timestamps,
     )
 
 
@@ -290,6 +338,7 @@ def transform_angular_velocity(
     timestamps=None,
     time_axis=None,
     cutoff=None,
+    return_timestamps=False,
 ):
     """ Transform an array of angular velocities between frames.
 
@@ -301,6 +350,18 @@ def transform_angular_velocity(
     another. In either case, it is assumed that the array is represented in
     the frame that is being changed and will be represented in the new
     frame after the transformation.
+
+    When transforming the reference frame R to a new frame R' while keeping
+    the moving frame M fixed, the transformed velocity is calculated
+    according to the formula:
+
+    .. math:: \omega_{M/R'} = \omega_{M/R} + \omega_{R/R'}
+
+    When transforming the moving frame M to a new frame M' while keeping
+    the reference frame R fixed, the transformed velocity is calculated
+    according to the formula:
+
+    .. math:: \omega_{M'/R} = \omega_{M/R} + \omega_{M'/M}
 
     Parameters
     ----------
@@ -344,6 +405,9 @@ def transform_angular_velocity(
         velocity after the twist estimation as a fraction of the Nyquist
         frequency.
 
+    return_timestamps: bool, default False
+        If True, also return the timestamps after the transformation.
+
     Returns
     -------
     arr_transformed: array_like
@@ -356,7 +420,7 @@ def transform_angular_velocity(
     --------
     transform_linear_velocity, transform_vectors, transform_quaternions,
     transform_points, ReferenceFrame
-    """
+    """  # noqa
     return _transform(
         "transform_angular_velocity",
         arr,
@@ -368,6 +432,7 @@ def transform_angular_velocity(
         time_axis,
         what=what,
         cutoff=cutoff,
+        return_timestamps=return_timestamps,
     )
 
 
@@ -384,6 +449,7 @@ def transform_linear_velocity(
     time_axis=None,
     cutoff=None,
     outlier_thresh=None,
+    return_timestamps=False,
 ):
     """ Transform an array of linear velocities between frames.
 
@@ -395,6 +461,18 @@ def transform_linear_velocity(
     another. In either case, it is assumed that the array is represented in
     the frame that is being changed and will be represented in the new
     frame after the transformation.
+
+    When transforming the reference frame R to a new frame R' while keeping
+    the moving frame M fixed, the transformed velocity is calculated
+    according to the formula:
+
+    .. math:: v_{M/R'} = v_{M/R} + v_{R/R'} + \omega_{R/R'} \\times t_{M/R}
+
+    When transforming the moving frame M to a new frame M' while keeping
+    the reference frame R fixed, the transformed velocity is calculated
+    according to the formula:
+
+    .. math:: v_{M'/R} = v_{M/R} + v_{M'/M} + \omega_{M/R} \\times t_{M'/M}
 
     Parameters
     ----------
@@ -456,6 +534,9 @@ def transform_linear_velocity(
         measurements from the Intel RealSense T265 tracker, set this value
         to 1e-3.
 
+    return_timestamps: bool, default False
+        If True, also return the timestamps after the transformation.
+
     Returns
     -------
     arr_transformed: array_like
@@ -468,7 +549,7 @@ def transform_linear_velocity(
     --------
     transform_angular_velocity, transform_vectors, transform_quaternions,
     transform_points, ReferenceFrame
-    """
+    """  # noqa
     return _transform(
         "transform_linear_velocity",
         arr,
@@ -483,6 +564,7 @@ def transform_linear_velocity(
         reference_frame=reference_frame,
         cutoff=cutoff,
         outlier_thresh=outlier_thresh,
+        return_timestamps=return_timestamps,
     )
 
 
@@ -567,8 +649,14 @@ def transform_coordinates(
         return arr
 
 
-def lookup_transform(outof, into, as_dataset=False):
+def lookup_transform(outof, into, as_dataset=False, return_timestamps=False):
     """ Look up transformation from one frame to another.
+
+    The transformation is a rotation `r` followed by a translation `t` which,
+    when applied to a point expressed wrt the base frame `B`, yields that
+    point wrt the target frame `T`:
+
+    .. math:: p_T = rot(r, p_B) + t
 
     Parameters
     ----------
@@ -580,43 +668,98 @@ def lookup_transform(outof, into, as_dataset=False):
 
     as_dataset: bool, default False
         If True, return an xarray.Dataset. Otherwise, return a tuple of
-        translation, rotation and timestamps.
+        translation and rotation.
+
+    return_timestamps: bool, default False
+        If True, and `as_dataset` is False, also return the timestamps of the
+        lookup.
 
     Returns
     -------
-    translation, rotation, timestamps: each numpy.ndarray
-        Translation, rotation and timestamps of transformation between the
-        frames, if `as_dataset` is False.
+    translation, rotation: each numpy.ndarray
+        Translation and rotation of transformation between the frames,
+        if `as_dataset` is False.
+
+    timestamps: numpy.ndarray
+        Corresponding timestamps of the lookup if `return_timestamps` is True.
 
     ds: xarray.Dataset
         The above arrays as an xarray.Dataset, if `as_dataset` is True.
     """
     into = _resolve_rf(into)
     outof = _resolve_rf(outof)
-    translation, rotation, timestamps = into.get_transformation(outof)
+    translation, rotation, timestamps = outof.lookup_transform(into)
 
     if as_dataset:
         return _make_transform_or_pose_dataset(
             translation, rotation, outof, timestamps
         )
-    else:
+    elif return_timestamps:
         return translation, rotation, timestamps
+    else:
+        return translation, rotation
+
+
+def lookup_pose(frame, reference, as_dataset=False, return_timestamps=False):
+    """ Look up pose of one frame wrt a reference.
+
+    Parameters
+    ----------
+    frame: str or ReferenceFrame
+        Frame for which to look up the pose.
+
+    reference: str or ReferenceFrame
+        Reference frame of the pose.
+
+    as_dataset: bool, default False
+        If True, return an xarray.Dataset. Otherwise, return a tuple of
+        position and orientation.
+
+    return_timestamps: bool, default False
+        If True, and `as_dataset` is False, also return the timestamps of the
+        lookup.
+
+    Returns
+    -------
+    position, orientation: each numpy.ndarray
+        Position and orientation of the pose between the frames,
+        if `as_dataset` is False.
+
+    timestamps: numpy.ndarray
+        Corresponding timestamps of the lookup if `return_timestamps` is True.
+
+    ds: xarray.Dataset
+        The above arrays as an xarray.Dataset, if `as_dataset` is True.
+    """
+    reference = _resolve_rf(reference)
+    frame = _resolve_rf(frame)
+    position, orientation, timestamps = frame.lookup_transform(reference)
+
+    if as_dataset:
+        return _make_transform_or_pose_dataset(
+            position, orientation, reference, timestamps, pose=True
+        )
+    elif return_timestamps:
+        return position, orientation, timestamps
+    else:
+        return position, orientation
 
 
 def lookup_twist(
-    moving_frame,
+    frame,
     reference=None,
     represent_in=None,
     outlier_thresh=None,
     cutoff=None,
     mode="quaternion",
     as_dataset=False,
+    return_timestamps=False,
 ):
     """ Estimate linear and angular velocity of a frame wrt a reference.
 
     Parameters
     ----------
-    moving_frame: str or ReferenceFrame
+    frame: str or ReferenceFrame
         The reference frame whose twist is estimated.
 
     reference: str or ReferenceFrame, optional
@@ -649,55 +792,65 @@ def lookup_twist(
 
     as_dataset: bool, default False
         If True, return an xarray.Dataset. Otherwise, return a tuple of linear
-        and angular velocity and timestamps.
+        and angular velocity.
+
+    return_timestamps: bool, default False
+        If True, and `as_dataset` is False, also return the timestamps of the
+        lookup.
 
     Returns
     -------
-    linear, angular, timestamps: each numpy.ndarray
-        Linear and angular velocity and timestamps of moving frame wrt
-        reference frame, represented in representation frame, if `as_dataset`
-        is False.
+    linear, angular: each numpy.ndarray
+        Linear and angular velocity of moving frame wrt reference frame,
+        represented in representation frame, if `as_dataset` is False.
+
+    timestamps: numpy.ndarray
+        Corresponding timestamps of the lookup if `return_timestamps` is True.
 
     ds: xarray.Dataset
         The above arrays as an xarray.Dataset, if `as_dataset` is True.
     """
-    moving_frame = _resolve_rf(moving_frame)
-    reference = _resolve_rf(reference or moving_frame.parent)
+    frame = _resolve_rf(frame)
+    reference = _resolve_rf(reference or frame.parent)
     represent_in = _resolve_rf(represent_in or reference)
 
-    linear, angular, timestamps = moving_frame.lookup_twist(
+    linear, angular, timestamps = frame.lookup_twist(
         reference,
         represent_in,
         outlier_thresh=outlier_thresh,
         cutoff=cutoff,
         mode=mode,
+        return_timestamps=True,
     )
 
     if as_dataset:
         return _make_twist_dataset(
-            angular, linear, moving_frame, reference, represent_in, timestamps
+            angular, linear, frame, reference, represent_in, timestamps
         )
-    else:
+    elif return_timestamps:
         return linear, angular, timestamps
+    else:
+        return linear, angular
 
 
 def lookup_linear_velocity(
-    moving_frame,
+    frame,
     reference=None,
     represent_in=None,
     outlier_thresh=None,
     cutoff=None,
     as_dataarray=False,
+    return_timestamps=False,
 ):
     """ Estimate linear velocity of a frame wrt a reference.
 
     Parameters
     ----------
-    moving_frame: str or ReferenceFrame
-        The reference frame whose twist is estimated.
+    frame: str or ReferenceFrame
+        The reference frame whose velocity is estimated.
 
     reference: str or ReferenceFrame, optional
-        The reference frame wrt which the twist is estimated. Defaults to
+        The reference frame wrt which the velocity is estimated. Defaults to
         the parent frame of the moving frame.
 
     represent_in: str or ReferenceFrame, optional
@@ -720,53 +873,62 @@ def lookup_linear_velocity(
         frequency.
 
     as_dataarray: bool, default False
-        If True, return an xarray.DataArray. Otherwise, return a tuple of
-        linear velocity and timestamps.
+        If True, return an xarray.DataArray.
+
+    return_timestamps: bool, default False
+        If True and `as_dataarray` is False, also return the timestamps of the
+        lookup.
 
     Returns
     -------
-    linear, timestamps: each numpy.ndarray
-        Linear velocity and timestamps of moving frame wrt
-        reference frame, represented in representation frame, if `as_dataarray`
-        is False.
+    linear: numpy.ndarray or xarray.DataArray
+        Linear velocity of moving frame wrt reference frame, represented in
+        representation frame.
 
-    da: xarray.DataArray
-        The above arrays as an xarray.DataArray, if `as_dataarray` is True.
+    timestamps: numpy.ndarray
+        Corresponding timestamps of the lookup if `return_timestamps` is True.
     """
-    moving_frame = _resolve_rf(moving_frame)
-    reference = _resolve_rf(reference or moving_frame.parent)
+    frame = _resolve_rf(frame)
+    reference = _resolve_rf(reference or frame.parent)
     represent_in = _resolve_rf(represent_in or reference)
 
-    linear, timestamps = moving_frame.lookup_linear_velocity(
-        reference, represent_in, outlier_thresh=outlier_thresh, cutoff=cutoff,
+    linear, timestamps = frame.lookup_linear_velocity(
+        reference,
+        represent_in,
+        outlier_thresh=outlier_thresh,
+        cutoff=cutoff,
+        return_timestamps=True,
     )
 
     if as_dataarray:
         return _make_velocity_dataarray(
-            linear, "linear", moving_frame, reference, represent_in, timestamps
+            linear, "linear", frame, reference, represent_in, timestamps
         )
-    else:
+    elif return_timestamps:
         return linear, timestamps
+    else:
+        return linear
 
 
 def lookup_angular_velocity(
-    moving_frame,
+    frame,
     reference=None,
     represent_in=None,
     outlier_thresh=None,
     cutoff=None,
     mode="quaternion",
     as_dataarray=False,
+    return_timestamps=False,
 ):
     """ Estimate angular velocity of a frame wrt a reference.
 
     Parameters
     ----------
-    moving_frame: str or ReferenceFrame
-        The reference frame whose twist is estimated.
+    frame: str or ReferenceFrame
+        The reference frame whose velocity is estimated.
 
     reference: str or ReferenceFrame, optional
-        The reference frame wrt which the twist is estimated. Defaults to
+        The reference frame wrt which the velocity is estimated. Defaults to
         the parent frame of the moving frame.
 
     represent_in: str or ReferenceFrame, optional
@@ -788,39 +950,39 @@ def lookup_angular_velocity(
         the gradient of the axis-angle representation of the rotations.
 
     as_dataarray: bool, default False
-        If True, return an xarray.DataArray. Otherwise, return a tuple of
-        angular velocity and timestamps.
+        If True, return an xarray.DataArray.
+
+    return_timestamps: bool, default False
+        If True and `as_dataarray` is False, also return the timestamps of the
+        lookup.
 
     Returns
     -------
-    angular, timestamps: each numpy.ndarray
-        Angular velocity and timestamps of moving frame wrt
-        reference frame, represented in representation frame, if `as_dataarray`
-        is False.
+    angular: numpy.ndarray or xarray.DataArray
+        Angular velocity of moving frame wrt reference frame, represented in
+        representation frame.
 
-    da: xarray.DataArray
-        The above arrays as an xarray.DataArray, if `as_dataarray` is True.
+    timestamps: numpy.ndarray
+        Corresponding timestamps of the lookup if `return_timestamps` is True.
     """
-    moving_frame = _resolve_rf(moving_frame)
-    reference = _resolve_rf(reference or moving_frame.parent)
+    frame = _resolve_rf(frame)
+    reference = _resolve_rf(reference or frame.parent)
     represent_in = _resolve_rf(represent_in or reference)
 
-    angular, timestamps = moving_frame.lookup_angular_velocity(
+    angular, timestamps = frame.lookup_angular_velocity(
         reference,
         represent_in,
         outlier_thresh=outlier_thresh,
         cutoff=cutoff,
         mode=mode,
+        return_timestamps=True,
     )
 
     if as_dataarray:
         return _make_velocity_dataarray(
-            angular,
-            "angular",
-            moving_frame,
-            reference,
-            represent_in,
-            timestamps,
+            angular, "angular", frame, reference, represent_in, timestamps,
         )
-    else:
+    elif return_timestamps:
         return angular, timestamps
+    else:
+        return angular