+++ tests/test_toplevel.py
@@ -163,13 +163,12 @@ class TestTopLevel:
             np.tile(arr_child2, (10, 1)),
             {"time": np.arange(10)},
             ("time", "cartesian_axis"),
-            attrs={"reference_frame": "child2"},
+            attrs={"representation_frame": "child2"},
         )
         da_child1 = rbm.transform_vectors(
             da_child2, into="child1", dim="cartesian_axis", timestamps="time",
         )
         assert da_child1.shape == (10, 3)
-        assert da_child1.attrs["reference_frame"] == "child1"
         assert da_child1.attrs["representation_frame"] == "child1"
         npt.assert_almost_equal(da_child1[0], arr_exp)
 
@@ -308,11 +307,13 @@ class TestTopLevel:
         xr = pytest.importorskip("xarray")
 
         # static frame
-        transform = rbm.lookup_transform("world", "child1", as_dataset=True)
+        transform = rbm.lookup_transform("world", "child2", as_dataset=True)
         assert set(transform.dims) == {"cartesian_axis", "quaternion_axis"}
-        np.testing.assert_equal(transform.translation.values, (1.0, 0.0, 0.0))
-        np.testing.assert_equal(
-            transform.rotation.values, (1.0, 0.0, 0.0, 0.0)
+        np.testing.assert_almost_equal(
+            transform.translation.values, (-1.0, 0.0, 0.0)
+        )
+        np.testing.assert_almost_equal(
+            transform.rotation.values, (0.0, 0.0, 0.0, -1.0)
         )
 
         # moving frame
@@ -321,17 +322,43 @@ class TestTopLevel:
             head_dataset, "position", "orientation", "time", "world", "head",
         ).register(update=True)
 
-        transform = rbm.lookup_transform("world", "head", as_dataset=True)
+        transform = rbm.lookup_transform("head", "world", as_dataset=True)
         xr.testing.assert_allclose(
             transform.translation, head_dataset.position
         )
-        assert transform.translation.attrs["reference_frame"] == "world"
-        assert transform.translation.attrs["representation_frame"] == "world"
+        assert transform.translation.attrs["reference_frame"] == "head"
+        assert transform.translation.attrs["representation_frame"] == "head"
         xr.testing.assert_allclose(
             transform.rotation, head_dataset.orientation
         )
-        assert transform.rotation.attrs["reference_frame"] == "world"
-        assert transform.rotation.attrs["representation_frame"] == "world"
+        assert transform.rotation.attrs["reference_frame"] == "head"
+        assert transform.rotation.attrs["representation_frame"] == "head"
+
+    def test_lookup_pose_xr(self, rf_tree, head_dataset):
+        """"""
+        xr = pytest.importorskip("xarray")
+
+        # static frame
+        pose = rbm.lookup_pose("world", "child2", as_dataset=True)
+        assert set(pose.dims) == {"cartesian_axis", "quaternion_axis"}
+        np.testing.assert_almost_equal(pose.position.values, (-1.0, 0.0, 0.0))
+        np.testing.assert_almost_equal(
+            pose.orientation.values, (0.0, 0.0, 0.0, -1.0)
+        )
+
+        # moving frame
+        rbm.register_frame("world", update=True)
+        rbm.ReferenceFrame.from_dataset(
+            head_dataset, "position", "orientation", "time", "world", "head",
+        ).register(update=True)
+
+        pose = rbm.lookup_pose("head", "world", as_dataset=True)
+        xr.testing.assert_allclose(pose.position, head_dataset.position)
+        assert pose.position.attrs["reference_frame"] == "world"
+        assert pose.position.attrs["representation_frame"] == "world"
+        xr.testing.assert_allclose(pose.orientation, head_dataset.orientation)
+        assert pose.orientation.attrs["reference_frame"] == "world"
+        assert pose.orientation.attrs["representation_frame"] == "world"
 
     def test_lookup_twist_xr(self, head_dataset):
         """"""